pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Amirali Qobadi <amiraliqobadi5@gmail.com>"]
license = "mit"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.3"
pillow = "^10.2.0"
django-taggit = "^5.0.1"
python-decouple = "^3.8"
python-dotenv = "^1.0.1"
django-admin-soft-dashboard = "^1.0.12"
django-jazzmin = "^3.0.0"
djangorestframework = "^3.15.1"
django-multiupload = "^0.6.1"
django-filter = "^24.2"
crispy-tailwind = "^1.0.3"
django-crispy-forms = "^2.1"
celery = "^5.4.0"
redis = "^5.0.4"
django-celery-beat = "^2.6.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.4"
pre-commit = "^3.7.0"
mixer = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
